+++ actionpack/lib/action_controller/metal/rendering.rb
@@ -98,11 +98,6 @@ module ActionController
           options[:html] = ERB::Util.html_escape(options[:html])
         end
 
-        if options.delete(:nothing)
-          ActiveSupport::Deprecation.warn("`:nothing` option is deprecated and will be removed in Rails 5.1. Use `head` method to respond with empty response body.")
-          options[:body] = nil
-        end
-
         if options[:status]
           options[:status] = Rack::Utils.status_code(options[:status])
         end